+++ trainer/downstream/train.py
@@ -137,7 +137,7 @@ class Trainer(comm_train.Trainer):
         for i, data in enumerate(loader):
             global_step = epoch * len(loader) + i + 1
             self.optimizer[ModelName.REPRESENTATIVE].zero_grad()
-            patch_image = data[DataLoaderKeys.IMAGE].to(self.device).float()
+            patch_image = data[DataLoaderKeys.IMAGE].to(self.device)
             _check_any_nan(patch_image)
             annot = data[DataLoaderKeys.LABEL].to(self.device).float()
 
@@ -369,7 +369,7 @@ class Trainer(comm_train.Trainer):
 
         for data in tqdm.tqdm(loader):
             # prediction
-            patch_image = data[DataLoaderKeys.IMAGE].to(self.device).float()
+            patch_image = data[DataLoaderKeys.IMAGE].to(self.device)
             _check_any_nan(patch_image)
 
             # annotation
